requirements/production.txt
coreapi==1.32.0
dj-database-url==0.4.1
Django==1.10
django-rest-swagger==0.3.10
djangorestframework==3.4.3
itypes==1.1.0
openapi-codec==1.0.0
prettyconf==1.2.3
psycopg2==2.6.2
requests==2.10.0
simplejson==3.8.2
uritemplate==0.6
whitenoise==3.2
unipath==1.1